Move addon tests to a new file
